web/Cheaters/web/requirements.txt
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.4
et-xmlfile==1.1.0
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
openpyxl==3.0.10
Pillow==9.3.0
pycparser==2.21
PyJWT==2.6.0
SQLAlchemy==1.4.44
uuid==1.30
Werkzeug==2.2.2
zipp==3.10.0
